docs/source/requirements.txt
numpydoc
matplotlib
sphinx>4.0.0   # https://github.com/readthedocs/readthedocs.org/issues/8616
sphinx_toggleprompt
furo
myst-parser
graphviz
